requirements.txt
pylint
requests
mysql-connector-python
python-dotenv